Add Version number constant to Host script, and add it to the UserAgent
Similar to the other plugin, we are using an Action to rewrite the version number
This can be switched to release number for releases by changing the release Action

This is in a script that is copied by the plugin when a host is added. This means it will be trivial for the customer to delete if we want.

I'm not sure where the architecture of this plugin is going next, but we will have much better control if we move to a runtime library and custom inspector